third_party/blink/web_tests/fast/css/inherit-initial-shorthand-values-expected.txt

PASS test('background') is "inherit, initial"
PASS test('background-position') is "inherit, initial"
PASS test('background-repeat') is "inherit, initial"
PASS test('border') is "inherit, initial"
PASS test('border-block-end') is "inherit, initial"
PASS test('border-block-start') is "inherit, initial"
PASS test('border-bottom') is "inherit, initial"
PASS test('border-color') is "inherit, initial"
PASS test('border-inline-end') is "inherit, initial"
PASS test('border-inline-start') is "inherit, initial"
PASS test('border-left') is "inherit, initial"
PASS test('border-radius') is "inherit, initial"
PASS test('border-right') is "inherit, initial"
PASS test('border-spacing') is "inherit, initial"
PASS test('border-style') is "inherit, initial"
PASS test('border-top') is "inherit, initial"
PASS test('border-width') is "inherit, initial"
PASS test('list-style') is "inherit, initial"
PASS test('font') is "inherit, initial"
PASS test('margin') is "inherit, initial"
PASS test('outline') is "inherit, initial"
PASS test('overflow') is "inherit, initial"
PASS test('padding') is "inherit, initial"
PASS test('webkit-animation') is "inherit, initial"
PASS test('webkit-columns') is "inherit, initial"
PASS test('webkit-column-rule') is "inherit, initial"
PASS test('webkit-flex-flow') is "inherit, initial"
PASS test('webkit-flex') is "inherit, initial"
PASS test('webkit-marginCollapse') is "inherit, initial"
PASS test('webkit-marquee') is "inherit, initial"
PASS test('webkit-mask') is "inherit, initial"
PASS test('webkit-mask-position') is "inherit, initial"
PASS test('webkit-mask-repeat') is "inherit, initial"
PASS test('webkit-text-emphasis') is "inherit, initial"
PASS test('webkit-text-stroke') is "inherit, initial"
PASS test('webkit-transition') is "inherit, initial"
PASS test('webkit-transform-origin') is "inherit, initial"
PASS successfullyParsed is true

TEST COMPLETE


Test that e matched declaration cache handles explicitly inherited properties correctly.
